Use apollo hooks for projects list
